role-based-access added

+++ app/accounts/columns.tsx
@@ -25,7 +25,8 @@ export const columns = (
   handleDelete: (account: z.infer<typeof ClientAccountSchema>) => void,
   handleSubAccount: (account: z.infer<typeof ClientAccountSchema>) => void,
   handleAddContact: (contact: z.infer<typeof ClientAccountSchema>) => void,
-  handleAssign: (account: z.infer<typeof ClientAccountSchema>) => void
+  handleAssign: (account: z.infer<typeof ClientAccountSchema>) => void,
+  userRole: "admin" | "consultant"
 ): ColumnDef<typeof ClientAccountSchema>[] => [
   {
     accessorKey: "id",
@@ -48,42 +49,43 @@ export const columns = (
     cell: ({ row }) => {
       const account = row.original;
 
-      return (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="ghost" className="h-8 w-8 p-0">
-              <span className="sr-only">Open menu</span>
-              <MoreHorizontal className="h-4 w-4" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <DropdownMenuLabel>Actions</DropdownMenuLabel>
-            <DropdownMenuItem onClick={() => handleEdit(account)}>
-              <Edit /> Edit account
-            </DropdownMenuItem>
-            <DropdownMenuItem
-              onClick={() => {
-                handleDelete(account);
-              }}
-            >
-              <Trash /> Delete account
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => handleSubAccount(account)}>
-              <Plus /> Add Sub-Account
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => handleAddContact(account)}>
-              <Plus /> Add Contacts
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => handleAssign(account)}>
-              <TestTube /> Assign consultants
-            </DropdownMenuItem>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem>
-              <Expand /> View details
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      );
+      if (userRole === "admin")
+        return (
+          <DropdownMenu>
+            <DropdownMenuTrigger asChild>
+              <Button variant="ghost" className="h-8 w-8 p-0">
+                <span className="sr-only">Open menu</span>
+                <MoreHorizontal className="h-4 w-4" />
+              </Button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent align="end">
+              <DropdownMenuLabel>Actions</DropdownMenuLabel>
+              <DropdownMenuItem onClick={() => handleEdit(account)}>
+                <Edit /> Edit account
+              </DropdownMenuItem>
+              <DropdownMenuItem
+                onClick={() => {
+                  handleDelete(account);
+                }}
+              >
+                <Trash /> Delete account
+              </DropdownMenuItem>
+              <DropdownMenuItem onClick={() => handleSubAccount(account)}>
+                <Plus /> Add Sub-Account
+              </DropdownMenuItem>
+              <DropdownMenuItem onClick={() => handleAddContact(account)}>
+                <Plus /> Add Contacts
+              </DropdownMenuItem>
+              <DropdownMenuItem onClick={() => handleAssign(account)}>
+                <TestTube /> Assign consultants
+              </DropdownMenuItem>
+              <DropdownMenuSeparator />
+              <DropdownMenuItem>
+                <Expand /> View details
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        );
     },
   },
 ];
